requirements.txt
absl-py==0.9.0
astor==0.8.1
astunparse==1.6.3
bokeh==2.1.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cloudpickle==1.5.0
cryptography==3.0
cycler==0.10.0
dask-glm==0.2.0
decorator==4.4.2
distributed==2.23.0
filelock==3.0.12
fsspec==0.8.0
gast==0.2.2
gdown==3.12.0
google-auth==1.20.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
HeapDict==1.0.1
idna==2.10
importlib-metadata==1.7.0
Jinja2==2.11.2
joblib==0.16.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
llvmlite==0.33.0
locket==0.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.0
msgpack==1.0.0
multipledispatch==0.6.0
numba==0.50.1
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.0
partd==1.1.0
Pillow==7.2.0
pkg-resources==0.0.0
protobuf==3.12.4
psutil==5.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-telegram-bot==12.8
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.4.1
six==1.15.0
sklearn==0.0
sortedcontainers==2.2.2
tblib==1.7.0
tensorboard==2.1.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
threadpoolctl==2.1.0
toolz==0.10.0
tornado==6.0.4
tqdm==4.48.0
typing-extensions==3.7.4.2
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zict==2.0.0
zipp==3.1.0
